backend/requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12.0
djangorestframework-simplejwt
psycopg2-binary
